fetch_function/requirements.txt
functions-framework==3.*
flask
psycopg2-binary
flask-cors